KM-17383: Migrate to SceneDelegate
